♻️refactor: files - 연관 주식 요약 탭 분리
#54
