+++ src/pages/main/RelatedStock.style.js
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+export const StyledMainContentDiv = styled.div`
+  width: 90%;
+  background-color: white;
+  display: flex;
+  flex-direction: column;
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2);
+  padding: 15px;
+  border-radius: 20px;
+  margin-bottom: 30px;
+  overflow-x: auto;
+  white-space: nowrap;
+`;
